requirements.txt
Pillow~=11.2.1


pytest~=8.4.1